start organization of API requests
